backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
razorpay>=1.4.0
Pillow>=10.0.0
gunicorn>=21.0.0
PyJWT>=2.8.0
